+++ lib/pacman-venv/activation/pacman-venv-activate.bash
@@ -22,8 +22,6 @@ exit() {
         unalias sudo
     fi
 
-    remove_shim_aliases
-
     # Unset all the temporary variables/functions
     unset _PACMAN_VENV_OLD_PATH
     unset _PACMAN_VENV_OLD_PS1
@@ -56,26 +54,10 @@ create_path() {
         new_path="${_PACMAN_VENV}${path}:${new_path}"
     done
 
+    # Add the shims directory to the beginning of the PATH
+    new_path="${_PACMAN_VENV}/pacman-venv-shims:${new_path}"
+
     echo "${new_path}"
-}
-
-# Point the package managers to Pacman Venv's command (include --root flag)
-# Need to be wrapped in single quotes (SC2139)
-create_shim_aliases() {
-    for shim in "${_PACMAN_VENV}"/pacman-venv-shims/*; do
-        alias_name="${shim##*.pacman-venv-}"
-        # shellcheck disable=SC2139
-        # These variables need to be expanded
-        alias "${alias_name}"="${shim}"
-    done
-}
-
-remove_shim_aliases() {
-    for shim in "${_PACMAN_VENV}"/pacman-venv-shims/*; do
-        alias_name="${shim##*.pacman-venv-}"
-        unalias "${alias_name}"
-    done
-
 }
 
 # Store all the old values of the variables
@@ -100,12 +82,10 @@ export PS1
 # Makepkg uses this variable to execute its Pacman command.
 # However, we need to override the default to use the --root flag,
 # so point makepkg to our pacman command.
-export PACMAN="${_PACMAN_VENV}/pacman-venv-shims/.pacman-venv-pacman"
+export PACMAN="${_PACMAN_VENV}/pacman-venv-shims/pacman"
 
 # Preserve env variables and allow sudo to be run with aliases
 alias sudo="sudo -E "
-
-create_shim_aliases
 
 # Forget past commands because the $PATH changed
 hash -r