+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Upcoming
 
+### Improvements
+- Using the shims in a script (i.e. `pacman`) will now use the virtual environment correctly
+
 ### Bug Fixes
 - Fix the man page not being installed in the correct location
+
 
 ## 0.1 (2023-04-23)